+++ hengband/src/main.cpp
@@ -286,6 +286,10 @@ int main(int argc, char *argv[])
 #endif /* SET_UID */
 
     auto browsing_movie = false;
+#ifdef USE_WEB
+    arg_sound = true;
+    arg_music = true;
+#endif
     for (auto i = 1; args && (i < argc); i++) {
         if (argv[i][0] != '-') {
             display_usage(argv[0]);